pyproject.toml
[project]
name = "ai-job-application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langgraph>=1.1.3",
    "npx>=0.1.6",
    "playwright>=1.58.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
]
